Clean up Opal scope
